+++ custom_components/trueachievements/__init__.py
@@ -2,22 +2,27 @@
 
 from __future__ import annotations
 
+import logging
+
 from typing import TYPE_CHECKING
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
-from .const import DOMAIN
+from .const import DOMAIN, STARTUP_MESSAGE
 from .coordinator import TrueAchievementsCoordinator
 
 if TYPE_CHECKING:
     from homeassistant.helpers.typing import ConfigType
 
 PLATFORMS: list[str] = ["binary_sensor", "sensor"]
+_LOGGER = logging.getLogger(__name__)
+
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up TrueAchievements from a config entry."""
+    _LOGGER.info(STARTUP_MESSAGE)
     session = async_get_clientsession(hass)
 
     coordinator = TrueAchievementsCoordinator(hass, entry, session)
@@ -32,9 +37,11 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
     return True
 
+
 async def update_listener(hass: HomeAssistant, entry: ConfigEntry) -> None:
     """Handle options update and reload the integration."""
     await hass.config_entries.async_reload(entry.entry_id)
+
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Unload a config entry."""